pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "numpy>=2.4.3",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "plotly>=6.6.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
]
